+++ src/app/api/galerie/albums/route.ts
@@ -1,55 +1,6 @@
-import { readdir, readFile } from "fs/promises";
+import { getListOfAlbumWithCoverImage } from "./getListOfAlbumWithCoverImage";
 
-export type ImageMetadata = {
-    album: string;
-    origin: string | null;
-    tags: string[];
-    description: string;
-    exif: { [key: string]: any } | null;
-    imgSize: {
-        width: number;
-        height: number;
-    };
-}
-
-export type ApiGaleryAlbumsResponse = Record<string, {
-    image: string;
-    fullsize: string;
-    metadata: ImageMetadata;
-}>;
-
-export async function getListOfAlbumWithCoverImage() {
-    // List all images in the public/Uploads directory
-    const basePath = "public";
-    const uploadsDir = "Galerie";
-    const thumbDir = "Galerie/thumb";
-    const images = (await readdir(`${basePath}/${uploadsDir}`, { withFileTypes: true })).sort((a, b) => a.name.localeCompare(b.name));
-
-    const filteredImageMetadatas = images.filter(
-        (file) => file.isFile() && file.name.endsWith(".json")
-    ).map(file => file.name);
-
-    // Each image should have a json file with metadata in the same directory
-    // If the image has no metadata, create a default metadata object
-    const imageData: ApiGaleryAlbumsResponse = {};
-
-    await Promise.all(filteredImageMetadatas.map(async (imageMeta) => {
-        const metadataPath = `${basePath}/${uploadsDir}/${imageMeta}`;
-        try {
-            const metadataFile = await readFile(metadataPath, "utf-8");
-            const metadata = JSON.parse(metadataFile);
-            imageData[metadata.album] = {
-                image: `/${thumbDir}/${imageMeta.replace(".json", "")}`,
-                fullsize: `/${uploadsDir}/${imageMeta.replace(".json", "")}`,
-                metadata,
-            };
-        } catch (error) {}
-    }));
-
-    return imageData;
-}
-
-export const GET = async (req: Request) => {
+export const GET = async () => {
     const imageData = await getListOfAlbumWithCoverImage();
 
     return new Response(JSON.stringify(imageData), {
