+++ src/app/api/galerie/albums/types.ts
@@ -1,0 +1,18 @@
+
+export type ImageMetadata = {
+    album: string;
+    origin: string | null;
+    tags: string[];
+    description: string;
+    exif: { [key: string]: unknown; } | null;
+    imgSize: {
+        width: number;
+        height: number;
+    };
+};
+
+export type ApiGaleryAlbumsResponse = Record<string, {
+    image: string;
+    fullsize: string;
+    metadata: ImageMetadata;
+}>;
